add client secure api
